# docs/openapi.yaml
openapi: 3.1.0
info:
  title: MYiTH Core API
  description: |
    채용공고와 국가직무능력표준(NCS)을 결합한 개인 맞춤형 취업 로드맵 서비스.

    인증: Google ID Token으로 로그인 후 발급받은 accessToken을
    Authorization: Bearer {token} 헤더에 실어 보낸다.

    공통 응답: 모든 성공 응답은 data로 감싼다. (GET /api/health 제외)
    공통 오류: { "error": { code, message, fieldErrors, requestId } }
  version: 0.1.0
servers:
- url: http://localhost:8080
  description: 로컬
- url: https://api.myith.example
  description: 운영
security:
- bearerAuth: []
...
paths:
...
  /api/uploads/presign:
    post:
      tags:
      - Upload
      summary: S3 Presigned URL 발급
      description: |-
        프론트가 uploadUrl로 S3에 직접 PUT 업로드하고, fileKey만 로드맵 생성에 넘긴다.
        서버는 파일 본문을 받지 않는다.
        허용 타입: application/pdf, image/png, image/jpeg. 상한 10MB.
      operationId: presign
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PresignRequest"
        required: true
      responses:
        "200":
          description: 발급 성공
          content:
            application/json:
              example:
                data:
                  uploadUrl: https://myith-uploads.s3.ap-northeast-2.amazonaws.com/...
                  fileKey: portfolio/usr_01J3ABC/9f2c1d.pdf
                  expiresIn: 900
        "400":
          description: UNSUPPORTED_FILE_TYPE / FILE_TOO_LARGE
          content:
            application/json:
              example:
                error:
                  code: UNSUPPORTED_FILE_TYPE
                  message: 허용되지 않는 파일 형식입니다.
                  requestId: req_01J3ABC
...
  /api/users/me:
    get:
      tags:
      - Auth
      summary: 내 정보 조회
      description: 현재 로그인한 사용자의 프로필 정보를 반환한다.
      operationId: getMe
      responses:
        "200":
          description: 조회 성공
          content:
            application/json:
              example:
                data:
                  id: usr_01J3ABC
                  email: sungyoon@example.com
                  nickname: 이성윤
                  profileImageUrl: null
                  createdAt: 2026-07-01T00:00:00Z
    delete:
      tags:
      - Auth
      summary: 회원 탈퇴
      description: "soft delete + PII 익명화. email→deleted_{id}@myith.local, google_id→\
        null, nickname→탈퇴한 사용자, profile_image_url→null."
      operationId: deleteMe
      responses:
        "204":
          description: 탈퇴 완료
    patch:
      tags:
      - Auth
      summary: 내 정보 수정
      description: 닉네임 또는 프로필 이미지를 변경한다. 모든 필드는 optional이다.
      operationId: updateMe
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMeRequest"
        required: true
      responses:
        "200":
          description: 수정 성공
          content:
            application/json:
              example:
                data:
                  id: usr_01J3ABC
                  email: sungyoon@example.com
                  nickname: 새이름
                  profileImageUrl: null
                  createdAt: 2026-07-01T00:00:00Z
        "422":
          description: VALIDATION_ERROR — nickname trim 후 1~20자
          content:
            application/json:
              example:
                error:
                  code: VALIDATION_ERROR
                  message: 입력값이 올바르지 않습니다.
                  fieldErrors:
                    nickname: 닉네임은 1~20자여야 합니다.
                  requestId: req_01J3ABC
...
  /api/star/records:
    get:
      tags:
      - Dashboard
      summary: 경험 카드(STAR 기록) 목록 조회
      description: 화면 5(대시보드) 경험 카드 목록. 커서 기반 페이지네이션(OFFSET 금지).
      operationId: getRecords
      parameters:
      - name: cursor
        in: query
        description: 커서
        required: false
        schema:
          type: string
        example: eyJpZCI6ImV4cF8wMSJ9
      - name: size
        in: query
        description: "페이지 크기. 기본 20, 최대 100"
        required: false
        schema:
          type: integer
          format: int32
          default: 20
        example: 20
      - name: axis
        in: query
        description: 축 코드로 필터. 생략 시 전체
        required: false
        schema:
          type: string
        example: programming
      - name: completeness
        in: query
        description: 완성도 필터. all | complete | partial (기본 all). complete = STAR 4칸
          모두 채워짐
        required: false
        schema:
          type: string
          enum:
          - all
          - complete
          - partial
      - name: tag
        in: query
        description: star_record.tags 배열 필터
        required: false
        schema:
          type: string
        example: 리더십
      responses:
        "200":
          description: 조회 성공
          content:
            application/json:
              example:
                data:
                - experienceId: exp_01
                  questId: qst_02
                  title: 언어 기초로 토이앱을 만든다
                  axisCode: programming
                  axisName: 프로그래밍 기초
                  ncsUnitName: 프로그래밍언어활용
                  star:
                    situation: 팀 프로젝트에서 API 응답이 3초 이상 걸리는 문제가 발생했다.
                    task: 응답 시간을 500ms 이하로 줄여야 했다.
                    action: N+1 쿼리를 페치 조인으로 개선하고 Redis 캐시를 도입했다.
                    result: 평균 응답 시간이 200ms로 감소했다.
                  createdAt: 2026-07-22T05:00:00Z
                meta:
                  nextCursor: eyJpZCI6ImV4cF8wMSJ9
                  hasNext: true
...
components:
  schemas:
...
    PresignRequest:
      type: object
      properties:
        fileName:
          type: string
          description: 업로드할 파일 이름
          example: portfolio.pdf
        contentType:
          type: string
          description: "MIME 타입. application/pdf, image/png, image/jpeg만 허용"
          example: application/pdf
      required:
      - contentType
      - fileName
...
    UpdateMeRequest:
      type: object
      properties:
        nickname:
          type: string
          description: 변경할 닉네임. trim 후 1~20자
          example: 새이름
          maxLength: 20
          minLength: 1
        profileImageUrl:
          type: string
          description: 변경할 프로필 이미지 URL
          example: https://example.com/avatar.png
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT